refactor: 각 경로의 에러 페이지를 글로벌 에러페이지로 대체
